Split up some tests for Travis
