+++ test/functional/commands/general/source-e2e-specs.js
@@ -2,8 +2,8 @@ import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import { DOMParser } from 'xmldom';
 import xpath from 'xpath';
-import { APIDEMOS_CAPS } from '../desired';
-import { initDriver } from '../helpers/session';
+import { APIDEMOS_CAPS } from '../../desired';
+import { initDriver } from '../../helpers/session';
 
 
 chai.should();
